+++ packages/e2e-tests/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- refactor `deploy.ts` to separate deploying test node / contracts / actors from setting up client
 - add subgraph E2E tests
 - Merkle behavior changed - it now only syncs up to the latest, non-empty committed leaf
 - Container names use uuid for randomness to avoid conflicts in CI